refactor: Move CO2 computation in metrics
