+++ src/lib/metrics.js
@@ -2,6 +2,21 @@ import { getSectionsInfo } from './trips'
 import {
   BICYCLING_MODE,
   WALKING_MODE,
+  AIR_MODE,
+  BUS_MODE,
+  CAR_MODE,
+  SUBWAY_MODE,
+  TRAIN_MODE,
+  UNKNOWN_MODE,
+  PLANE_CO2_KG_PER_KM_SHORT,
+  PLANE_CO2_KG_PER_KM_MEDIUM,
+  PLANE_CO2_KG_PER_KM_LONG,
+  SHORT_PLANE_TRIP_MAX_DISTANCE,
+  MEDIUM_PLANE_TRIP_MAX_DISTANCE,
+  CAR_AVERAGE_CO2_KG_PER_KM,
+  BUS_MEDIUM_AGGLOMERATION_CO2_KG_PER_KM,
+  TRAIN_HIGHLINE_CO2_KG_PER_KM,
+  SUBWAY_TRAM_TROLLEYBUS_MEDIUM_AGGLOMERATION_CO2_KG_PER_KM,
   DEFAULT_WEIGHT_KG,
   SLOW_WALKING_MAX_SPEED,
   MEDIUM_WALKING_MAX_SPEED,
@@ -18,6 +33,56 @@ import {
   SLOW_BICYCLING_MAX_MEDIUM,
   SLOW_BICYCLING_MAX_FAST
 } from 'src/constants/const'
+
+/**
+ * Compute the total CO2 consumed by the trip.
+ * For each section, compute the related consumed CO2 based on the mode and distance.
+ * See the src/constants/const.js file for more insights about the CO2 values
+ *
+ * @param {object} trip - The GeoJSON trip
+ * @returns {number} The consumed CO2, in kg
+ */
+export const computeCO2Trip = trip => {
+  const sections = getSectionsInfo(trip)
+  let totalCO2 = 0
+  for (const section of sections) {
+    const distance = section.distance / 1000 // convert in km
+    switch (section.mode) {
+      case AIR_MODE:
+        if (distance <= SHORT_PLANE_TRIP_MAX_DISTANCE) {
+          totalCO2 += distance * PLANE_CO2_KG_PER_KM_SHORT
+          break
+        } else if (distance <= MEDIUM_PLANE_TRIP_MAX_DISTANCE) {
+          totalCO2 += distance * PLANE_CO2_KG_PER_KM_MEDIUM
+        } else {
+          totalCO2 += distance * PLANE_CO2_KG_PER_KM_LONG
+        }
+        break
+      case CAR_MODE:
+        // TODO: should depends on the energy type + number of passengers
+        totalCO2 += distance * CAR_AVERAGE_CO2_KG_PER_KM
+        break
+      case BUS_MODE:
+        // TODO: should depends on the area
+        totalCO2 += distance * BUS_MEDIUM_AGGLOMERATION_CO2_KG_PER_KM
+        break
+      case SUBWAY_MODE:
+        // TODO: should depends on the area
+        totalCO2 +=
+          distance * SUBWAY_TRAM_TROLLEYBUS_MEDIUM_AGGLOMERATION_CO2_KG_PER_KM
+        break
+      case TRAIN_MODE:
+        // TODO: should depends on train type
+        totalCO2 += distance * TRAIN_HIGHLINE_CO2_KG_PER_KM
+        break
+      case UNKNOWN_MODE:
+        break
+      default:
+        break
+    }
+  }
+  return totalCO2
+}
 
 /**
  * Compute the calories produced based on weight, MET and duration.
